+++ apps/workflows-dashboard/src/pages/Workflows/components/organisms/metrics/ActivePerWorkflow/hooks/useActivePerWorkflowStatsQuery/useActivePerWorkflowStatsQuery.ts
@@ -3,6 +3,7 @@ import { useQuery } from '@tanstack/react-query';
 
 export const useActivePerWorkflowStatsQuery = () => {
   const { data = [], isLoading, isFetching } = useQuery(workflowMetricsKeys.workflowRuntimeStats());
+  console.log('data', data);
 
   return {
     data,