+++ sdks/web-ui-sdk/CHANGELOG.md
@@ -1,5 +1,12 @@
 # web-ui-sdk
 
+## 1.3.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.5.4
+
 ## 1.3.4
 
 ### Patch Changes