+++ services/workflows-service/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @ballerine/workflows-service
 
+## 0.4.15
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.5.4
+
 ## 0.4.14
 
 ### Patch Changes